+++ international_online_offer/views.py
@@ -44,7 +44,7 @@ class IOOIntent(FormView):
 class IOOLocation(FormView):
     form_class = forms.LocationForm
     template_name = 'ioo/triage/location.html'
-    success_url = reverse_lazy('international_online_offer:index')
+    success_url = reverse_lazy('international_online_offer:hiring')
 
     def get_context_data(self, **kwargs):
         return super().get_context_data(
@@ -54,4 +54,36 @@ class IOOLocation(FormView):
             question_text='Where in the UK would you like to expand your business?',
             why_we_ask_this_question_text="""We'll use this information to provide customised content
               relevant to your city, county or region.""",
+        )
+
+
+class IOOHiring(FormView):
+    form_class = forms.HiringForm
+    template_name = 'ioo/triage/hiring.html'
+    success_url = reverse_lazy('international_online_offer:spend')
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(
+            **kwargs,
+            back_url='international_online_offer:location',
+            step_text='Step 4 of 5',
+            question_text='How many people are you looking to hire in the UK?',
+            why_we_ask_this_question_text="""We'll use this information to provide customised content
+              relevant to your hiring.""",
+        )
+
+
+class IOOSpend(FormView):
+    form_class = forms.SpendForm
+    template_name = 'ioo/triage/spend.html'
+    success_url = reverse_lazy('international_online_offer:index')
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(
+            **kwargs,
+            back_url='international_online_offer:hiring',
+            step_text='Step 5 of 5',
+            question_text='What is your planned spend for UK entry or expansion?',
+            why_we_ask_this_question_text="""We'll use this information to provide customised content
+              relevant to your spend.""",
         )